requirements.txt
Django
djangorestframework
django-allauth
celery
redis
psycopg2-binary
django-cors-headers
django-celery-beat
django-celery-results
django-filter
Pillow
requests
whitenoise